Added GitHub actions with custom images
Compiling gRPC takes about 8 minutes, which is better to do in advance.
The docker images (for php 7.2 - 8.0) contains grpc and some other
extensions required for running basic unit tests, and skip this waiting
time. This allows for a huge time save, instead of
( 8 min * laravel versions * php version * releases ).

Compiling of a single image takes about 15 minutes, so I added a manual
job to compile the images and publish them to the GitHub image repo.

Note: I tried compiling them on Alpine first, but that gave far to many
errors.